update metrics-core dependency to latest version, refactored Influxdb to be an interface rather than a class, move old Influxdb class to InfludbHttp, and add InfluxdbUdp class
