+++ pom.xml
@@ -32,10 +32,9 @@
 	</licenses>
 	<dependencies>
 		<dependency>
-			<groupId>com.codahale.metrics</groupId>
+			<groupId>io.dropwizard.metrics</groupId>
 			<artifactId>metrics-core</artifactId>
-			<version>3.0.1</version>
-			<scope>provided</scope>
+			<version>3.1.0</version>
 		</dependency>
 		<dependency>
 			<groupId>org.slf4j</groupId>
